backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
google-cloud-aiplatform==1.71.1
google-cloud-firestore==2.16.0
vertexai==1.71.1
python-dotenv==1.0.1
pydantic==2.7.1
